cockpit.cpp: Fix a silly warning

+++ engine/src/gfx/cockpit.cpp
@@ -945,7 +945,7 @@ void GameCockpit::Delete() {
         delete text;
         text = nullptr;
     }
-    for (size_t i = 0; i < (int) mesh.size(); ++i) {
+    for (size_t i = 0; i < mesh.size(); ++i) {
         if (mesh[i] != nullptr) {
             delete mesh[i];
             mesh[i] = nullptr;
